\name{LocLab}
 
\alias{LocLab}

\title{Function for better position of the labels in the graphs.}

\description{Function for better position of the labels in the graphs.}

\usage{
LocLab(x, y = NULL, labels = seq(along = x), cex = 1,
       method = c("SANN", "GA"), allowSmallOverlap = FALSE,
       trace = FALSE, shadotext = FALSE, 
       doPlot = TRUE, \dots)
}

\arguments{
  \item{x}{Coordinate x}
  \item{y}{Coordinate y}
  \item{labels}{The labels}
  \item{cex}{cex}
  \item{method}{Not used}
  \item{allowSmallOverlap}{Boolean}
  \item{trace}{Boolean}
  \item{shadotext}{Boolean}
  \item{doPlot}{Boolean}
  \item{\dots}{Other arguments passed to or from other methods}
}

\value{
See the text of the function.
}

